Refactoreing app state into stores
